Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.11
pydantic-settings==2.5.2
PyJWT==2.9.0
google-auth==2.35.0
python-dotenv==1.0.1
python-multipart==0.0.12
alembic==1.13.3
requests==2.31.0